requirements.txt
Babel==2.9.1  # min version = 2.6.0 (Focal with security backports)

chardet==3.0.4

cryptography==2.6.1; python_version < '3.12'  # incompatibility between pyopenssl 19.0.0 and cryptography>=37.0.0

cryptography==42.0.8 ; python_version >= '3.12'  # (Noble) min 41.0.7, pinning 42.0.8 for security fixes

decorator==4.4.2

docutils==0.16

ebaysdk==2.1.5

freezegun==0.3.11; python_version < '3.8'

freezegun==0.3.15; python_version >= '3.8'

idna==2.8

Jinja2==2.11.3 ; python_version < '3.12' # min version = 2.10.1 (Focal - with security backports)

Jinja2==3.1.2 ; python_version >= '3.12'  # (Noble) compatibility with markupsafe

libsass==0.18.0 ; python_version < '3.12'

libsass==0.22.0 ; python_version >= '3.12'  # (Noble) Mostly to have a wheel package

lxml==4.6.5; python_version < '3.12' # min version = 4.5.0 (Focal - with security backports)

lxml==5.2.1; python_version >= '3.12' # (Noble - removed html clean)

lxml-html-clean; python_version >= '3.12' # (Noble - removed from lxml, unpinned for futur security patches)

MarkupSafe==1.1.0 ; python_version < '3.12'

MarkupSafe==2.1.5 ; python_version >= '3.12'  # (Noble) Mostly to have a wheel package

num2words==0.5.6

ofxparse==0.19; python_version <= '3.9'

ofxparse==0.21; python_version > '3.9'  # (Jammy)

passlib==1.7.3 # min version = 1.7.2 (Focal with security backports)

Pillow==9.0.1 ; python_version < '3.12' # min version = 7.0.0 (Focal with security backports)

Pillow==10.2.0 ; python_version >= '3.12'  # (Noble) Mostly to have a wheel package

polib==1.1.0

psutil==5.6.7 ; python_version < '3.12' # min version = 5.5.1 (Focal with security backports)

psutil==5.9.8 ; python_version >= '3.12'  # (Noble) Mostly to have a wheel package



pydot==1.4.1

pyopenssl==19.0.0 ; python_version < '3.12'

pyopenssl==24.1.0 ; python_version >= '3.12' # (Noble) min 23.2.0, pinned for compatibility with cryptography==42.0.8 and security patches

PyPDF2==1.26.0 ; python_version < '3.12'

PyPDF2==2.12.1 ; python_version >= '3.12' # (Noble) Compatibility with cryptography

pypiwin32 ; sys_platform == 'win32'

pyserial==3.4

python-dateutil==2.8.2

python-ldap==3.4.0 ; sys_platform != 'win32' and python_version < '3.12' # min version = 3.2.0 (Focal with security backports)

python-ldap==3.4.4 ; sys_platform != 'win32' and python_version >= '3.12'  # (Noble) Mostly to have a wheel package

python-stdnum==1.13

pytz  # no version pinning to avoid OS perturbations

pyusb==1.0.2 ; python_version < '3.12'

pyusb==1.2.1 ; python_version >= '3.12'  # (Noble) Mostly to have a wheel package

qrcode==6.1

reportlab==3.5.59 ; python_version < '3.12' # version < 3.5.54 are not compatible with Pillow 8.1.2 and 3.5.59 is bullseye

reportlab==4.1.0 ; python_version >= '3.12' # (Noble) Mostly to have a wheel package

requests==2.25.1 ;  python_version < '3.12' # versions < 2.25 aren't compatible w/ urllib3 1.26. Bullseye = 2.25.1. min version = 2.22.0 (Focal)

requests==2.31.0 ; python_version >= '3.12' # (Noble) Compatibility with i

rl-renderPM==4.0.3 ; sys_platform == 'win32' and python_version >= '3.12'  # Needed by reportlab 4.1.0 but included in deb package

urllib3==1.26.5 ;  python_version < '3.12' # indirect / min version = 1.25.8 (Focal with security backports)

urllib3==2.0.7  ; python_version >= '3.12'  # (Noble) Compatibility with cryptography

vobject==0.9.6.1

Werkzeug==0.16.1 ; python_version <= '3.9'

Werkzeug==2.0.2 ; python_version > '3.9'  # (Jammy)

xlrd==1.1.0; python_version < '3.8'

xlrd==1.2.0; python_version >= '3.8'

XlsxWriter==1.1.2

xlwt==1.3.0

zeep==3.4.0

pandas==2.2.2

paramiko==2.12.0

psycopg2-binary

pdfminer.six

# AI & LLM Integration for Document Summarization
anthropic==0.25.6  # Claude API
openai==1.50.0  # ChatGPT API (fixed version)
google-genai>=1.47.0  # Google Gemini API (NEW: replaces google-generativeai)
google-generativeai==0.5.0  # Google Gemini API (OLD: kept for compatibility)
google-auth-httplib2==0.2.0  # Google Cloud Authentication
google-auth-oauthlib==1.2.0  # Google OAuth
google-api-python-client==2.100.0  # Google Drive API
